fix: move sanitizeReturnTo to @/utils/ and use it in review-terms
The inline returnTo validation missed backslash-based open redirects.
Move sanitizeReturnTo from @/domains/account/ to @/utils/return-to so
it can be imported across domains without violating the cross-domain
lint rule. Re-export from the old location for existing consumers.
